Added serverless-dynamodb-client which creates a dynamodb singleton for sharing across lambdas.
It's important to use this, otherwise each lambda will create it's own dynamodb db instance and the serverless schema will not be preserved. Updated all handlers to make use of the serverless-dynamodb-client. Updated the schema to use the title as the primary key. Updated the README to use movie examples. Added movies.json & load-movies.js for allowing easy dynamodb data upload. Renamed schema.js to config.js and added a schema endpoint for identify the current dynamodb table schema - this helped debug the singleton instance issue.